refactor: extract types, predicates, selectors from usePayloadSyncFilters
456 lines mixed filter types and sentinel constants, pure predicates,
pure list/stat derivation, and the React state machine that drives the
country -> city -> neighborhood cascade.

Extract into three modules under features/admin/utils, leaving the hook
with state, memos, and the four auto-reset effects (456 -> 299 lines):
- payload-sync-filter.types.ts    types + UNSPECIFIED_* sentinels
- payload-sync-filter-predicates.ts  matchesStatusFilter, formatLabel, facets
- payload-sync-selectors.ts       list building, filter composition, stats

The three derived lists (hierarchy base, type base, filtered) each
repeated the same status/category/scope/type filtering inline. They now
compose shared selectors, which is where a subtle behavior change could
hide, so this was verified differentially rather than by reading: a
harness ran the original inline memo bodies against the new composition
over every combination of the six filters (24,193 comparisons, 400
generated locations) plus the stats block. All outputs identical.

The hook re-exports the constants, types, formatLabel and
matchesStatusFilter it used to define, so all 5 importing modules are
untouched. Returned object keys unchanged (25/25).

Verified: tsc -p tsconfig.app.json stays at 0 errors.
